Move hint_utils to hint_processor
