+++ src/hint_processor/builtin_hint_processor/builtin_hint_processor.rs
@@ -4,6 +4,11 @@ use std::collections::HashMap;
 use num_bigint::BigInt;
 
 use crate::any_box;
+use crate::hint_processor::hint_processor_definition::{HintReference, HintProcessor};
+use crate::serde::deserialize_program::ApTracking;
+use crate::types::exec_scope::ExecutionScopesProxy;
+use crate::types::instruction::Register;
+use crate::vm::errors::vm_errors::VirtualMachineError;
 use crate::hint_processor::builtin_hint_processor::blake2s_utils::{
     blake2s_add_uint256, blake2s_add_uint256_bigend, compute_blake2s, finalize_blake2s,
 };
@@ -11,20 +16,14 @@ use crate::hint_processor::builtin_hint_processor::dict_hint_utils::{
     default_dict_new, dict_new, dict_read, dict_squash_copy_dict, dict_squash_update_ptr,
     dict_update, dict_write,
 };
-use crate::hint_processor::builtin_hint_processor::find_element_hint::{
-    find_element, search_sorted_lower,
-};
+use crate::hint_processor::builtin_hint_processor::find_element_hint::{find_element, search_sorted_lower};
 use crate::hint_processor::builtin_hint_processor::hint_code;
-use crate::hint_processor::builtin_hint_processor::keccak_utils::{
-    unsafe_keccak, unsafe_keccak_finalize,
-};
-use crate::hint_processor::builtin_hint_processor::math_utils::*;
-use crate::hint_processor::builtin_hint_processor::memcpy_hint_utils::{
+use crate::hint_processor::hint_utils::{
     add_segment, enter_scope, exit_scope, memcpy_continue_copying, memcpy_enter_scope,
 };
-use crate::hint_processor::builtin_hint_processor::memset_utils::{
-    memset_continue_loop, memset_enter_scope,
-};
+use crate::hint_processor::builtin_hint_processor::keccak_utils::{unsafe_keccak, unsafe_keccak_finalize};
+use crate::hint_processor::builtin_hint_processor::math_utils::*;
+use crate::hint_processor::builtin_hint_processor::memset_utils::{memset_continue_loop, memset_enter_scope};
 use crate::hint_processor::builtin_hint_processor::pow_utils::pow;
 use crate::hint_processor::builtin_hint_processor::set::set_add;
 use crate::hint_processor::builtin_hint_processor::squash_dict_utils::{
@@ -36,12 +35,6 @@ use crate::hint_processor::builtin_hint_processor::squash_dict_utils::{
 use crate::hint_processor::builtin_hint_processor::uint256_utils::{
     split_64, uint256_add, uint256_signed_nn, uint256_sqrt, uint256_unsigned_div_rem,
 };
-use crate::hint_processor::hint_processor_definition::{HintProcessor, HintReference};
-use crate::hint_processor::hint_utils::bigint_to_usize;
-use crate::serde::deserialize_program::ApTracking;
-use crate::types::exec_scope::ExecutionScopesProxy;
-use crate::types::instruction::Register;
-use crate::vm::errors::vm_errors::VirtualMachineError;
 
 use crate::hint_processor::builtin_hint_processor::cairo_keccak::keccak_hints::{
     block_permutation, cairo_keccak_finalize, compare_bytes_in_word_nondet,
@@ -58,15 +51,15 @@ use crate::hint_processor::builtin_hint_processor::secp::{
     },
     signature::{div_mod_n_packed_divmod, div_mod_n_safe_div, get_point_from_x},
 };
-use crate::hint_processor::builtin_hint_processor::sha256_utils::{
-    sha256_finalize, sha256_input, sha256_main,
-};
+use crate::hint_processor::builtin_hint_processor::sha256_utils::{sha256_finalize, sha256_input, sha256_main};
 use crate::hint_processor::builtin_hint_processor::usort::{
     usort_body, usort_enter_scope, verify_multiplicity_assert, verify_multiplicity_body,
     verify_usort,
 };
 use crate::vm::vm_core::{VMProxy, VirtualMachine};
 use crate::vm::vm_memory::memory::get_memory_proxy;
+
+use crate::hint_processor::hint_utils::bigint_to_usize;
 
 pub struct HintProcessorData {
     pub code: String,
