+++ src/hint_processor/builtin_hint_processor/cairo_keccak/keccak_hints.rs
@@ -1,6 +1,6 @@
-use crate::hint_processor::builtin_hint_processor::hint_utils::get_integer_from_var_name;
-use crate::hint_processor::builtin_hint_processor::hint_utils::get_ptr_from_var_name;
-use crate::hint_processor::builtin_hint_processor::hint_utils::insert_value_into_ap;
+use crate::hint_processor::hint_utils::get_integer_from_var_name;
+use crate::hint_processor::hint_utils::get_ptr_from_var_name;
+use crate::hint_processor::hint_utils::insert_value_into_ap;
 use crate::{
     bigint,
     hint_processor::hint_processor_definition::HintReference,
