+++ src/hint_processor/builtin_hint_processor/find_element_hint.rs
@@ -1,17 +1,16 @@
-use crate::bigint;
-use crate::hint_processor::builtin_hint_processor::hint_utils::get_relocatable_from_var_name;
+use crate::hint_processor::hint_processor_definition::HintReference;
+use crate::hint_processor::hint_utils::{
+    bigint_to_usize, get_ptr_from_var_name, get_relocatable_from_var_name,
+    insert_value_from_var_name,
+};
 use crate::serde::deserialize_program::ApTracking;
 use crate::types::exec_scope::ExecutionScopesProxy;
 use crate::vm::errors::vm_errors::VirtualMachineError;
 use crate::vm::vm_core::VMProxy;
+use crate::{bigint, hint_processor::hint_utils::get_integer_from_var_name};
 use num_bigint::BigInt;
 use num_traits::{Signed, ToPrimitive};
 use std::collections::HashMap;
-
-use super::hint_utils::get_ptr_from_var_name;
-use super::hint_utils::insert_value_from_var_name;
-use super::hint_utils::{bigint_to_usize, get_integer_from_var_name};
-use crate::hint_processor::hint_processor_definition::HintReference;
 
 pub fn find_element(
     vm_proxy: &mut VMProxy,
