+++ src/hint_processor/builtin_hint_processor/keccak_utils.rs
@@ -1,7 +1,7 @@
-use super::hint_utils::{
+use crate::hint_processor::hint_processor_definition::HintReference;
+use crate::hint_processor::hint_utils::{
     get_integer_from_var_name, get_ptr_from_var_name, get_relocatable_from_var_name,
 };
-use crate::hint_processor::hint_processor_definition::HintReference;
 use crate::{
     bigint,
     serde::deserialize_program::ApTracking,
