+++ src/hint_processor/builtin_hint_processor/math_utils.rs
@@ -8,7 +8,7 @@ use num_bigint::BigInt;
 use num_integer::Integer;
 use num_traits::{Signed, Zero};
 
-use super::hint_utils::{
+use crate::hint_processor::hint_utils::{
     get_address_from_var_name, get_integer_from_var_name, get_ptr_from_var_name,
     get_range_check_builtin, insert_value_from_var_name, insert_value_into_ap,
 };
