+++ src/hint_processor/builtin_hint_processor/mod.rs
@@ -6,9 +6,9 @@ pub mod dict_hint_utils;
 pub mod dict_manager;
 pub mod find_element_hint;
 pub mod hint_code;
-pub mod hint_utils;
 pub mod keccak_utils;
 pub mod math_utils;
+pub mod memcpy_hint_utils;
 pub mod memset_utils;
 pub mod pow_utils;
 pub mod secp;
